requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
Pillow>=10.0.0